[RENAME] 헤더 폴더 구조 변경
- common 폴더 내 header 폴더 넣기
- Header 컴포넌트를 HeaderLeft와 두 개로 분리
- 다른 파일들은 Header 경로 바뀌면서 바뀜
