+++ kakaobase/src/app/login/page.tsx
@@ -1,4 +1,4 @@
-import Header from '@/components/common/Header';
+import Header from '@/components/common/header/Header';
 import LoginForm from '@/components/user/LoginForm';
 
 export default function Page() {
